+++ providers/wikidata.ts
@@ -4,22 +4,43 @@ import { Points, LngLat } from '../utils'
 const wikidataCodes = require('./wikidata/codes.json')
 const wikidataLanguages: Array<string> = require('./wikidata/languages.json')
 
+export const Options: Options = {
+  subclasses: ['Q486972'],
+  languages: ['en', 'fr', 'es', 'de', 'it', 'ru'],
+  radius: 15,
+}
 export interface Options extends utils.Options {
-  limit?: number
   nearest?: LngLat
   radius?: number
   subclasses?: Array<string>
   languages?: Array<string>
 }
-export interface Result { }
-export interface Results { }
 
-export function createQuery(address: string, options: Options) {
+interface Entity {
+  type: string
+  value: string
+  datatype?: string
+  'xml:lang'?: string
+}
+
+export interface Result {
+  place?: Entity
+  location?: Entity
+  distance?: Entity
+  placeDescription?: Entity
+  [lanugage: string]: Entity
+}
+export interface Results {
+  head: { vars: Array<string> }
+  results: { bindings: Array<Result> }
+}
+
+export function createQuery(address: string, options?: Options) {
   // Options
   const [lng, lat] = options.nearest
-  const radius = options.radius || 15
-  const subclasses = options.subclasses || ['Q486972']
-  const languages = options.languages || ['en', 'fr', 'es', 'de', 'it', 'ru']
+  const radius = options.radius || Options.radius
+  const subclasses = options.subclasses || Options.subclasses
+  const languages = options.languages || Options.languages
 
   // Validate languages
   languages.map(language => {
@@ -53,33 +74,57 @@ export function createQuery(address: string, options: Options) {
   }
     `
   }
-  query += `
-  # Filter by Exact Name
-  OPTIONAL {
-`
+  query += `\n  # Filter by Exact Name\n`
   languages.map(language => {
-    query += `    ?place rdfs:label ?name_${ language } FILTER (lang(?name_${ language }) = "${ language }") .\n`
+    query += `  OPTIONAL {?place rdfs:label ?name_${ language } FILTER (lang(?name_${ language }) = "${ language }") . }\n`
   })
 
-  query += `  }\n\n`
-  query += ` FILTER (`
+  query += `\n  FILTER (`
   query += languages.map(language => `regex(?name_${ language }, "^${ address }$")`).join(' || ')
-  query += `) .`
+  query += `) .\n`
 
   // Descriptions
   query += `
-# Get Descriptions
+  # Get Descriptions
   SERVICE wikibase:label {
     bd:serviceParam wikibase:language "${ languages.join(',') }"
   }
 
 } ORDER BY ASC(?dist)`
+  return query
 }
 
 /**
  * Convert Wikidata SPARQL results into GeoJSON
  */
 export function toGeoJSON(json: Results, options?: Options): Points {
+  const languages = options.languages || Options.languages
   const collection: Points = turf.featureCollection([])
+  json.results.bindings.map(result => {
+    // Standard Wikidata tags
+    const id = result.place.value.match(/entity\/(.+)/)[1]
+    const [lng, lat] = result.location.value.match(/\(([\-\.\d]+) ([\-\.\d]+)\)/).slice(1, 3).map(n => Number(n))
+    const distance = Number(result.distance.value)
+    const description = result.placeDescription.value
+    const properties: any = {
+      id,
+      distance,
+      description,
+    }
+    // Parse languages
+    languages.map(language => {
+      const match = result[`name_${ language }`]
+      if (match !== undefined) {
+        properties[`name:${ language }`] = match.value
+      }
+    })
+
+    // Create Point
+    const point = turf.point([lng, lat], properties)
+    point.id = id
+
+    // Add to GeoJSON Feature Collection
+    collection.features.push(point)
+  })
   return collection
 }